fix: 共通部分をcoreファイルに移動
