+++ core/types.ts
@@ -1,0 +1,4 @@
+export interface Error {
+  message: string;
+  errorCode: number;
+}
