+++ v1/game.ts
@@ -1,12 +1,12 @@
 import { Router } from "../deps.ts";
 
 import { contentTypeFilter, jsonParse } from "./util.ts";
-import { accounts, kkmm, tournaments } from "./datas.ts";
-import { errors, ServerError } from "./error.ts";
+import { accounts, kkmm, tournaments } from "../core/datas.ts";
+import { errors, ServerError } from "../core/error.ts";
 import { Game as IGame, GameCreateReq } from "./types.ts";
 import { auth } from "./middleware.ts";
-import { getAllBoards, getBoard } from "./parts/firestore_opration.ts";
-import { ExpGame } from "./parts/expKakomimasu.ts";
+import { getAllBoards, getBoard } from "../core/firestore_opration.ts";
+import { ExpGame } from "../core/expKakomimasu.ts";
 
 export const gameRouter = () => {
   const router = new Router();
