+++ v1/match.ts
@@ -2,8 +2,8 @@ import { Core, Router } from "../deps.ts";
 
 import { nowUnixTime } from "../core/util.ts";
 import { contentTypeFilter, jsonParse } from "./util.ts";
-import { accounts, kkmm } from "./datas.ts";
-import { errors, ServerError } from "./error.ts";
+import { accounts, kkmm } from "../core/datas.ts";
+import { errors, ServerError } from "../core/error.ts";
 import {
   ActionPost as IActionPost,
   ActionReq,
@@ -14,9 +14,9 @@ import {
 } from "./types.ts";
 import { auth } from "./middleware.ts";
 import { aiList } from "./parts/ai-list.ts";
-import { nonReqEnv } from "./parts/env.ts";
-import { ExpGame, Player } from "./parts/expKakomimasu.ts";
-import { getAllBoards, getBoard } from "./parts/firestore_opration.ts";
+import { nonReqEnv } from "../core/env.ts";
+import { ExpGame, Player } from "../core/expKakomimasu.ts";
+import { getAllBoards, getBoard } from "../core/firestore_opration.ts";
 
 const boardname = nonReqEnv.boardname; // || "E-1"; // "F-1" "A-1"
