+++ v1/middleware.ts
@@ -1,7 +1,7 @@
 import { Context } from "../deps.ts";
 
-import { accounts } from "./datas.ts";
-import { errorCodeResponse, errors, ServerError } from "./error.ts";
+import { accounts } from "../core/datas.ts";
+import { errorCodeResponse, errors, ServerError } from "../core/error.ts";
 import { getPayload } from "./parts/jwt.ts";
 
 export const auth = (
