--- v1/parts/interface.ts
@@ -1,7 +1,0 @@
-export interface IBoard {
-  name: string;
-  width: number;
-  height: number;
-  nagent: number;
-  points: number[];
-}
