+++ v1/util.ts
@@ -1,5 +1,5 @@
 import { Context } from "../deps.ts";
-import { errors, ServerError } from "./error.ts";
+import { errors, ServerError } from "../core/error.ts";
 
 export type UnknownRequest<T> = Record<keyof T, unknown>;
