+++ v1/ws.ts
@@ -2,8 +2,8 @@ import { Router } from "../deps.ts";
 
 import type { WsGameReq, WsGameRes } from "./types.ts";
 import { UnknownRequest } from "./util.ts";
-import { accounts, kkmm } from "./datas.ts";
-import { ExpGame } from "./parts/expKakomimasu.ts";
+import { accounts, kkmm } from "../core/datas.ts";
+import { ExpGame } from "../core/expKakomimasu.ts";
 
 type SearchOptions = { op: string; value: string }[];
 type MapValue = {
